Flatten implementation
